OEP, solver, tests; close all findings
Round-3 multi-agent audit produced 4 reports
(reports_local/audit_round3/{A_physics_constraints,C_solvers_oneshot,D_oep,F_test_quality}.md).
This commit clears every finding flagged in those reports plus the R1/R2
carryovers explicitly listed by as still open.

oep.py
L1: delete dead h = mf.get_hcore() in _baseline_vxc_matrix
L5: symmetrize three_center against AO-quadrature noise so
V_xc^baseline + Σ b_t · 3c[t] is exactly Hermitian in (i,j)
L6: append ' + final_scf_failed' to lbfgs_status when the
post-optimization SCF blows up; downstream consumers can
now distinguish scipy failure from final-SCF failure
L7: defensive np.asarray + 3-D normalization on j_matrix in
_ks_from_vxc_matrix_uhf for PySCF-version robustness
L3: document n_iter clip in OEPResult docstring

losses.py
NEW N1: drop dtype=jnp.float64 from the four zero-fallback
returns; consistent with N6's pbe_anchor fix and
lets x32 callers avoid silent dtype promotion

solver.py
N3: add CRITERION_REGISTRY + register_criterion decorator,
mirroring MIXER_REGISTRY; EnergyConvergence now self-registers
N4: DEGENERACY_REG, SYM_BREAK_SHIFT, _GOLDEN_RATIO and
_sym_break_diag now live here as the single source of
truth across all SCF backends

solver_manual.py
N3: _build_criterion now consults CRITERION_REGISTRY
N4: re-import shared constants from solver
R1 M8 : mixer_state frozen on the already-converged branch via
jax.tree_util.tree_map; pre-fix code advanced step_index
(and any future history-tracking mixer's state) every
cycle past convergence

solver_pyscfad.py
N1: JIT-tracer guard scans ALL likely-traced keys (dm_pbe,
j_matrix, rho_grid, ao_grid, s_matrix, h_core), not just
the first present one
R1 H4 : _features_for_block zero-pads the trailing rows when
pyscfad's block_loop emits a non-uniform block (last
block, non0tab pruning); padded positions have zero
grid weight downstream so they contribute nothing to
the Fock/energy sums
N5: drop dtype=jnp.float64 pin from trace_arr
R1 M9 : replace dead getattr(config, 'max_cycles', ...) with
direct field access; SolverConfig always defines it
R1 L5 : _holder_get_veff threads caller-supplied mol_ through
both _reassemble_features_on_grid and original_get_veff
instead of substituting the closed-over mol

oneshot.py
N4: re-export shared constants/utility from solver

tests
@pytest.mark.slow on
test_precompute_passes_spin_resolved_dm_for_uks (runs full
PBE SCF on CH triplet)
test_losses.py : rewrite test_dm_term_normalizes_per_element_for_uks
to actually call _dm_term with a monkey-patched prediction
stub; the prior version re-derived an arithmetic identity
and would still pass under a regression to n_ao**2 divisor
test_oep.py : strengthen test_oep_baseline_xc_parameter_accepts_arbitrary_xc
with a behavioral discriminator (V_xc differs across LDA/PBE/BLYP
baselines) plus 'hf' coverage; rename
test_oep_v_space_regularization_basis_independent →
test_oep_v_space_regularization_uses_aux_overlap to match
what the test actually checks

regression sweep: 498 passed across 5 active chunks of the alec test
suite (chunk 6 collected zero non-slow tests, expected). No new
warnings or xfail regressions vs the post-R2 baseline.

+++ xcquinox/alec/losses.py
@@ -236,7 +236,7 @@ def _dm_term(model, mol_data, iter_idx, solver_config=None, relative=False):
             n_elems = int(jnp.prod(jnp.array(dm_ref_arr.shape)))
             err = err / float(n_elems)
         terms.append(err)
-    return jnp.mean(jnp.stack(terms)) if terms else jnp.array(0.0, dtype=jnp.float64)
+    return jnp.mean(jnp.stack(terms)) if terms else jnp.array(0.0)
 
 
 def _grid_term(model, mol_data, iter_idx, solver_config=None, relative=False):
@@ -252,7 +252,7 @@ def _grid_term(model, mol_data, iter_idx, solver_config=None, relative=False):
         if relative:
             err = err / (jnp.sum(w * rho_ref ** 2) + 1e-8)
         terms.append(err)
-    return jnp.mean(jnp.stack(terms)) if terms else jnp.array(0.0, dtype=jnp.float64)
+    return jnp.mean(jnp.stack(terms)) if terms else jnp.array(0.0)
 
 
 def _vxc_term(model, mol_data, iter_idx, relative=False):
@@ -304,13 +304,13 @@ def _vxc_term(model, mol_data, iter_idx, relative=False):
                 n_ao = vxc_ref_arr.shape[-1]
                 err = err / (n_ao * n_ao)
         terms.append(err)
-    return jnp.mean(jnp.stack(terms)) if terms else jnp.array(0.0, dtype=jnp.float64)
+    return jnp.mean(jnp.stack(terms)) if terms else jnp.array(0.0)
 
 
 def _anchor_term(model, sample, weight: float) -> jnp.ndarray:
     """PBE-anchor loss: weight * mean((F_x_nn - F_x_PBE)^2) on a fixed sample."""
     if sample is None or weight == 0.0:
-        return jnp.array(0.0, dtype=jnp.float64)
+        return jnp.array(0.0)
     from xcquinox.alec.pbe_anchor import pbe_anchor_loss
     from xcquinox.alec.oneshot import _nn_fx_local_uks
     def _nn_fx(m, rho_alpha, rho_beta, s_vals):
